+++ packages/perseus-editor/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @khanacademy/perseus-editor
 
+## 28.15.0
+
+### Minor Changes
+
+-   [#3310](https://github.com/Khan/perseus/pull/3310) [`1ed915b32b`](https://github.com/Khan/perseus/commit/1ed915b32b279518d35de9517ed33b497ebc5468) Thanks [@nishasy](https://github.com/nishasy)! - [Image] | (cleanup) | Remove image-widget-upgrade-alignment feature flag
+
+*   [#3314](https://github.com/Khan/perseus/pull/3314) [`a734a7479c`](https://github.com/Khan/perseus/commit/a734a7479cd32c02fc427b467406b12f56081669) Thanks [@nishasy](https://github.com/nishasy)! - [Image] | (CX) | Disable image sizing fields
+
+### Patch Changes
+
+-   Updated dependencies [[`1ed915b32b`](https://github.com/Khan/perseus/commit/1ed915b32b279518d35de9517ed33b497ebc5468)]:
+    -   @khanacademy/perseus@75.7.0
+    -   @khanacademy/perseus-core@23.4.0
+    -   @khanacademy/keypad-context@3.2.36
+    -   @khanacademy/kmath@2.2.36
+    -   @khanacademy/math-input@26.4.5
+    -   @khanacademy/perseus-linter@4.8.4
+    -   @khanacademy/perseus-score@8.2.13
+
+## 28.14.4
+
+### Patch Changes
+
+-   [#3308](https://github.com/Khan/perseus/pull/3308) [`250ae305fe`](https://github.com/Khan/perseus/commit/250ae305fe86f1f0ea0e7b9061ab1410969af20e) Thanks [@nishasy](https://github.com/nishasy)! - [Image] | (CX) | Add InfoTip explaining alignment info to Article editor
+
+*   [#3305](https://github.com/Khan/perseus/pull/3305) [`810a8d419c`](https://github.com/Khan/perseus/commit/810a8d419ca937a5682753f548fcbebbfa368479) Thanks [@nishasy](https://github.com/nishasy)! - [Image] | (a11y) | Update alt text limit to 125 characters
+
+*   Updated dependencies [[`810a8d419c`](https://github.com/Khan/perseus/commit/810a8d419ca937a5682753f548fcbebbfa368479), [`e181ce906d`](https://github.com/Khan/perseus/commit/e181ce906d54a2515efa5a7338f7e627223acffe), [`5159e994de`](https://github.com/Khan/perseus/commit/5159e994de7ba010f7838ae7f71ccff4c5ac9832), [`83402a8bf5`](https://github.com/Khan/perseus/commit/83402a8bf5e89153e7681ea03adb77b50e6c8975), [`e108dc754d`](https://github.com/Khan/perseus/commit/e108dc754df0b78cb44763875d76fa8958564e17)]:
+    -   @khanacademy/perseus-linter@4.8.3
+    -   @khanacademy/perseus@75.6.0
+    -   @khanacademy/perseus-core@23.3.0
+    -   @khanacademy/keypad-context@3.2.35
+    -   @khanacademy/kmath@2.2.35
+    -   @khanacademy/math-input@26.4.4
+    -   @khanacademy/perseus-score@8.2.12
+
+## 28.14.3
+
+### Patch Changes
+
+-   Updated dependencies [[`77404ad02e`](https://github.com/Khan/perseus/commit/77404ad02e0647c9b1a90c3aa0b162d834f4d183)]:
+    -   @khanacademy/perseus@75.5.0
+
 ## 28.14.2
 
 ### Patch Changes