+++ packages/perseus-linter/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/perseus-linter
 
+## 4.8.4
+
+### Patch Changes
+
+-   Updated dependencies [[`1ed915b32b`](https://github.com/Khan/perseus/commit/1ed915b32b279518d35de9517ed33b497ebc5468)]:
+    -   @khanacademy/perseus-core@23.4.0
+    -   @khanacademy/kmath@2.2.36
+
+## 4.8.3
+
+### Patch Changes
+
+-   [#3305](https://github.com/Khan/perseus/pull/3305) [`810a8d419c`](https://github.com/Khan/perseus/commit/810a8d419ca937a5682753f548fcbebbfa368479) Thanks [@nishasy](https://github.com/nishasy)! - [Image] | (a11y) | Update alt text limit to 125 characters
+
+-   Updated dependencies [[`83402a8bf5`](https://github.com/Khan/perseus/commit/83402a8bf5e89153e7681ea03adb77b50e6c8975)]:
+    -   @khanacademy/perseus-core@23.3.0
+    -   @khanacademy/kmath@2.2.35
+
 ## 4.8.2
 
 ### Patch Changes